# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rezervacija resursa",
        "description": "API za login, registraciju i upravljanje korisnicima",
        "termsOfService": "",
        "contact": {
            "email": "mislav.zabcic@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "reservations/request": {
            "post": {
                "tags": [
                    "reservations"
                ],
                "summary": "Zatra\u017ei opremu",
                "description": "Zatra\u017ei opremu",
                "operationId": "api.reservations.request",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "body",
                        "description": "Datum po\u010detka posudbe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "return_date",
                        "in": "body",
                        "description": "Datum vra\u0107anja opreme, null ukoliko ne zna\u0161 kolko ti treba",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "body",
                        "description": "Ime",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "remark",
                        "in": "body",
                        "description": "Napomena uz posudbu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request made"
                    },
                    "409": {
                        "description": "User exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Registracija",
                "description": "Registracija",
                "operationId": "api.auth.register",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Email adresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Lozinka",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "first_name",
                        "in": "body",
                        "description": "Ime",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "body",
                        "description": "Prezime",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created"
                    },
                    "409": {
                        "description": "User exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
        "auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Prijava",
                "description": "Prijava",
                "operationId": "api.auth.login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Korisni\u010dko ime za prijavu u sustav",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Lozinka",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login succesfull",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid password",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "404": {
                        "description": "User doesn't exist'",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
        "users/current": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Prika\u017ei trenutnog korisnika",
                "description": "Prika\u017ei koji si korisnik",
                "operationId": "api.users.current",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current user object",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
    },
    "definitions": {
        "CustomError": {
            "properties": {
                "error": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
...
        "LoginResponse": {
            "properties": {
                "token": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "user": {
                    "description": "",
                    "format": "User",
                    "$ref": "#/definitions/User"
                }
            },
            "type": "object"
        },
...
        "User": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int"
                },
                "first_name": {
                    "type": "integer",
                    "format": "string"
                },
                "last_name": {
                    "type": "string",
                    "format": "string"
                },
                "email": {
                    "type": "string",
                    "format": "string"
                },
                "role_id": {
                    "type": "integer",
                    "format": "int"
                },
                "created_at": {
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        }
    },
...
}